pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-skill"
version = "0.1.0"
description = "Convert any MCP server into an Agent Skill"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = ["mcp", "agent", "skill", "code-generation", "fastmcp", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastmcp>=2.0",
    "click>=8.0",
    "jinja2>=3.1",
    "ruff>=0.15.4",
    "ty>=0.0.20",
    "py-key-value-aio[disk,filetree]",
    "asyncclick>=8.3.0.3",
]

[project.urls]
Homepage = "https://github.com/manojbajaj95/mcp-skill"
Repository = "https://github.com/manojbajaj95/mcp-skill"
Issues = "https://github.com/manojbajaj95/mcp-skill/issues"

[project.scripts]
mcp-skill = "mcp_skill.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_skill"]
